pypeit image
